// veolms_backend/src/modules/payments/zod/payment.schema.ts
import { z } from "zod"

export const createOrderSchema = z.object({
    courseId: z.string().trim().min(1),
})

export const verifyPaymentSchema = z.object({
    razorpay_order_id: z.string().trim().min(1),
    razorpay_payment_id: z.string().trim().min(1),
    razorpay_signature: z.string().trim().min(1),
})